Move hook to a separate file
