+++ src/components/Biscuit/RawBiscuit/hook.js
@@ -1,0 +1,16 @@
+const useBiscuit = (onAnimationEnd, onStamp) => {
+  let animationsCount = 0;
+
+  const handleAnimationEnd = () => {
+    animationsCount++;
+
+    if (animationsCount % 2 === 0) {
+      onAnimationEnd(true);
+      onStamp(false);
+    }
+  };
+
+  return { handleAnimationEnd };
+};
+
+export default useBiscuit;
