+++ src/components/Biscuit/RawBiscuit/index.js
@@ -1,19 +1,6 @@
 import "./style.css";
 
-const useBiscuit = (onAnimationEnd, onStamp) => {
-  let animationsCount = 0;
-
-  const handleAnimationEnd = () => {
-    animationsCount++;
-
-    if (animationsCount % 2 === 0) {
-      onAnimationEnd(true);
-      onStamp(false);
-    }
-  };
-
-  return { handleAnimationEnd };
-};
+import useBiscuit from "./hook.js";
 
 const Biscuit = ({ onAnimationEnd, state, onStamp }) => {
   const { handleAnimationEnd } = useBiscuit(onAnimationEnd, onStamp);
